requirements.txt
asgiref==3.8.1
Django==5.0.7
sqlparse==0.5.1
typing_extensions==4.12.2
drf-spectacular==0.27.2
psycopg==3.2.1
psycopg-binary==3.2.1
pandas
scikit-learn
faiss-cpu
transformers
